Struct rsp2_potentials::crespi::Params

pub struct Params {
    pub delta: f64,
    pub lambda: f64,
    pub A: f64,
    pub z0: f64,
    pub C: f64,
    pub C2N: [f64; 3],
    pub cutoff_begin: f64,
    pub cutoff_transition_dist: Option<f64>,
}

Fields
delta: f64

Transverse distance scaling factor. Units are Angstroms.

lambda: f64

Distance scaling factor for the repulsive potential. Units are inverse Angstroms.

A: f64

Multiplicative factor for the attractive potential. Units are eV.

z0: f64

“Convenience scaling factor.” Units are Angstroms.

C: f64

Offset constant used in repulsive potential. Units are eV.

C2N: [f64; 3]

Multiplicative constants used in repulsive potential. Contains C0, C2, C4. Units are eV.

cutoff_begin: f64

Distance at which the Kolmogorov-Crespi potential starts to switch to zero. Units are Angstroms.

cutoff_transition_dist: Option<f64>

How far it takes for the Kolmogorov-Crespi to switch to zero after the cutoff distance. Units are Angstroms.

A value of None can be used to simulate Lammps’ behavior of a sharp cutoff. In this case, the value will be offset slightly for C0 continuity.

Implementations

impl Params

pub fn original() -> Params

Constants used for calculation of the Kolmogorov-Crespi potential.

These match the values used by default by the implementation of kolmogorov/crespi/z in Lammps, which is scaled to Lammps’ rebo’s bond length.

Citation

A.N. Kolmogorov & V. H. Crespi, Registry-dependent interlayer potential for graphitic systems. Physical Review B 71, 235415 (2005)

pub fn ouyang() -> Params

Constants used for calculation of the Kolmogorov-Crespi potential.

These match the values used by default by the implementation of kolmogorov/crespi/full in Lammps.

Citation

Wengen Ouyang, Davide Mandelli, Michael Urbakh, Oded Hod, arXiv:1806.09555 (2018).

impl Params

pub fn compute(&self, r_ij: V3, normal_i: V3, normal_j: V3) -> Output

Calculate the Kolmogorov-Crespi potential.

Accepts the delta vector between two carbon atoms r_ij as well as the unit normal vectors for each atom, n_i and n_j. Returns the potential as well as the gradient of the potential with respect to r_ij, n_i, and n_j.

For the total potential of a system, one should only sum over undirected bonds. (see FracBond::is_canonical).

pub fn compute_z(&self, r_ij: V3) -> (f64, V3)

Computes crespi assuming normals are all +Z.

pub fn compute_z_with_hessian(&self, r_ij: V3) -> (f64, V3, M33)

Computes crespi assuming normals are all +Z.

impl Params

pub fn cutoff_end(&self) -> f64

Distance after which the Kolmogorov-Crespi potential is always zero.

pub fn value_offset(&self) -> f64

Offset added to value for C0 continuity at the cutoff.

This is zero if a smooth cutoff is enabled (cutoff_transition_dist).