Trait rsp2_potentials::interactions::Potential

pub trait Potential: Sync {
    type AtomType: Clone + Sync;
    fn check_distance(
        &self, 
        distance: f64, 
        types: (Self::AtomType, Self::AtomType)
    ) -> FailResult<IsInteracting>;

    fn max_bonds() -> Option<usize>;
}

The trait a potential needs to implement to use Interactions.

Associated Types

type AtomType: Clone + Sync

Loading content...
Required methods

fn check_distance(
&self,
distance: f64,
types: (Self::AtomType, Self::AtomType)
) -> FailResult<IsInteracting>

Determine whether a bond of this length should be included in the bond list.

Some potentials may choose to fail for some bond lengths. (e.g. potentials that can’t handle reactions may fail on borderline values of length). Some implementations may also choose to log warnings.

fn max_bonds() -> Option<usize>

Loading content...