Module rsp2_potentials::rebo::nonreactive

Implementation of the second-generation REBO potential, without the reactive bits

This implementation was written with the help of the AIREBO paper, and by reviewing the implementation in LAMMPS.

Reactive parts

The code was originally written to support fractional weights, but this was dropped for ease of maintenance, since rsp2 does not currently actually require a reactive potential. If bond lengths in the transition radius are found, rsp2 will bail out.

Dropping support for fractional weights significantly simplifies the computation of forces.

The original, almost certainly VERY buggy reactive code still exists in the git history (see 557b47c570517b8d604ce9767937fdf225084c14 and diff reactive.rs and nonreactive.rs), and many parts of the current code are still written in a manner that will make it less difficult to add them back by carrying them around and multiplying them in even though they are all 1.0 or 0.0.

Citations

2nd gen REBO: Donald W Brenner et al 2002 J. Phys.: Condens. Matter 14 783

AIREBO: Steven J Stuart et al J. Chem. Phys. 112, 6472 (2000)

LAMMPS: S. Plimpton, J Comp Phys, 117, 1-19 (1995)

Optimized parameters for phonons: L. Lindsay and D. A. Broido, Phys. Rev. B, 81, 205441 (2010)

Structs

BondGrad

InteractionsPotential

Params

Enums

AtomType

Constants

SITE_MAX_BONDS
The most bonds an atom can have.

Functions

compute

compute_by_bond

find_all_interactions

Type Definitions

Interactions

SiteBondVec
A stack-allocated vector of data whose indices correspond to the elements returned by interactions.bonds(site) for some particular site, in order.